requirements.txt
9==9
absl-py==0.13.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
anyio==3.2.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bayesian-optimization==1.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
blessed==1.18.0
blessings==1.7
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
captum==0.4.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cma==3.1.0
colorama==0.4.4
contextvars==2.4
cupy-cuda110 @ file:///home/phuocbui/MCMC/reimplement/cupy_cuda110-9.0.0-cp36-cp36m-manylinux1_x86_64.whl
cycler==0.10.0
Cython==0.29.23
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.9.1
dash-table==4.12.0
dask==2021.3.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.1
Django==3.0.14
dowhy==0.6
dtaidistance==2.2.5
entrypoints==0.3
et-xmlfile==1.1.0
fastrlock==0.6
filelock==3.0.12
finta==1.3
Flask==2.0.1
Flask-Compress==1.10.1
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
google-api-core==1.31.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
idna==3.2
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.6.1
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.10.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.2
multidict==5.1.0
multitasking==0.0.9
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nevergrad==0.4.3.post4
notebook==6.4.0
numba==0.53.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2020.11.15
plotly==5.1.0
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.9.1
ptyprocess==0.7.0
py-spy==0.3.7
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl @ git+https://github.com/pyro-ppl/pyro.git@9f67c43deab7f33216b426843292e5c5ce73df78
pyrsistent==0.18.0
pyspark==3.1.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
ray==1.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.7.1
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
statsmodels==0.12.2
sympy==1.8
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tomlkit==0.7.0
toolz==0.11.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.6
virtualenv==20.4.6
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.4
yarl==1.6.3
yfinance==0.1.63
zipp==3.5.0
